Rip find_closest_with_paths out of XG and make it an algorithm
